Fix "count" command.
Fixed the "count" command. Before this fix, the last saved message count would be added to the current counter even before the message that marked that count was reached. This opened the possibility for the count to be inaccurate, because if the last saved message was deleted, it would never be reached, so the current counter would continue running even after "reaching" the last saved message, and the saved message count would be added to the current counter. E.g.: If a server had 1000 messages and they were all deleted, when running the "count" command again, those 1000 messages would also be added to the current counter, causing inconsistency in the actual value.