Fix a possible error in "count" command.
This is a simple fix, but prior to it, users could theoretically use the "count" command when passing a message ID or referencing it, then delete said message after the bot counted the amount of messages up to it, but before it could reply to it. This would cause an exception which would be ignored, and the bot could incorrectly update the message count on the database.